backend/requirements/dev.txt
asgiref==3.9.1
Django==5.2.6
django-environ==0.12.0
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.3
asgiref==3.9.1
channels==4.2.2
channels_redis==4.3.0
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
msgpack==1.1.1
packaging==25.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
sqlparse==0.5.3
uritemplate==4.2.0
